Numerous minor changes, including:
Migration of variables to CCI

Insertion of FIXME and TODO flags for future work

Beginning of efforts to add input and output validation to functions

this is the first revision where the main compiles properly.  Many of the peripheral libraries will not yet compile without significant error, more work is required on those.