Numerous minor changes, including:
Migration of variables to CCI

Insertion of FIXME and TODO flags for future work

Beginning of efforts to add input and output validation to functions

Modification of some variables to constants, as they do not change throughout execution, and verification that any value referenced by a pointer is declared static as to prevent dereferencing errors.

this is the first revision where the main compiles properly.  Many of the peripheral libraries will not yet compile without significant error, more work is required on those.